move run_cmd_stream_output to its own class
